+++ FcmSharp/FcmSharp.Test/Integration/BatchMessageIntegrationTest.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.IO;
-using System.Linq;
 using System.Threading;
 using FcmSharp.Requests;
 using FcmSharp.Settings;
@@ -25,27 +24,21 @@ namespace FcmSharp.Test.Integration
             // Construct the Client:
             using (var client = new FcmClient(settings))
             {
-                var notification = new Notification
-                {
-                    Title = "Notification Title",
-                    Body = "Notification Body Text"
-                };
 
-                var messages = tokens
-                    // Create the Messages to be sent for each of the Tokens:
-                    .Select(token => new Message
+                Message message = new Message
+                {
+                    Notification = new Notification
                     {
-                        Token = token,
-                        Notification = notification
-                    })
-                    // And turn it into an Array:
-                    .ToArray();
+                        Title = "Notification Title",
+                        Body = "Notification Body Text"
+                    }
+                };
 
                 // Finally send the Message and wait for the Result:
                 CancellationTokenSource cts = new CancellationTokenSource();
 
                 // Send the Message and wait synchronously:
-                var result = client.SendBatchAsync(messages, false, cts.Token).GetAwaiter().GetResult();
+                var result = client.SendMulticastMessage(tokens, message, false, cts.Token).GetAwaiter().GetResult();
 
                 // Print the Result to the Console:
                 foreach (var response in result.Responses)